+++ remote-wallet/src/lib.rs
@@ -1,6 +1,8 @@
 #![cfg(feature = "agave-unstable-api")]
 #![allow(clippy::arithmetic_side_effects)]
 #![allow(dead_code)]
+#[cfg(feature = "keystone")]
+pub mod keystone;
 pub mod ledger;
 pub mod ledger_error;
 pub mod locator;