+++ remote-wallet/src/locator.rs
@@ -14,11 +14,15 @@ pub enum Manufacturer {
     Unknown,
     Ledger,
     Trezor,
+    #[cfg(feature = "keystone")]
+    Keystone,
 }
 
 const MANUFACTURER_UNKNOWN: &str = "unknown";
 const MANUFACTURER_LEDGER: &str = "ledger";
 const MANUFACTURER_TREZOR: &str = "trezor";
+#[cfg(feature = "keystone")]
+const MANUFACTURER_KEYSTONE: &str = "keystone";
 
 #[derive(Clone, Debug, Error, PartialEq, Eq)]
 #[error("not a manufacturer")]
@@ -37,6 +41,8 @@ impl FromStr for Manufacturer {
         match s.as_str() {
             MANUFACTURER_LEDGER => Ok(Self::Ledger),
             MANUFACTURER_TREZOR => Ok(Self::Trezor),
+            #[cfg(feature = "keystone")]
+            MANUFACTURER_KEYSTONE => Ok(Self::Keystone),
             _ => Err(ManufacturerError),
         }
     }
@@ -55,6 +61,8 @@ impl AsRef<str> for Manufacturer {
             Self::Unknown => MANUFACTURER_UNKNOWN,
             Self::Ledger => MANUFACTURER_LEDGER,
             Self::Trezor => MANUFACTURER_TREZOR,
+            #[cfg(feature = "keystone")]
+            Self::Keystone => MANUFACTURER_KEYSTONE,
         }
     }
 }