+++ remote-wallet/src/remote_keypair.rs
@@ -3,8 +3,7 @@ use {
         ledger::get_wallet_from_info,
         locator::Locator,
         remote_wallet::{
-            RemoteWallet, RemoteWalletError, RemoteWalletInfo, RemoteWalletManager,
-            RemoteWalletType,
+            RemoteWalletError, RemoteWalletInfo, RemoteWalletManager, RemoteWalletType,
         },
     },
     solana_derivation_path::DerivationPath,
@@ -27,10 +26,7 @@ impl RemoteKeypair {
         confirm_key: bool,
         path: String,
     ) -> Result<Self, RemoteWalletError> {
-        let pubkey = match &wallet_type {
-            RemoteWalletType::Ledger(wallet) => wallet.get_pubkey(&derivation_path, confirm_key)?,
-            RemoteWalletType::Trezor(wallet) => wallet.get_pubkey(&derivation_path, confirm_key)?,
-        };
+        let pubkey = wallet_type.get_pubkey(&derivation_path, confirm_key)?;
 
         Ok(Self {
             wallet_type,
@@ -47,14 +43,9 @@ impl Signer for RemoteKeypair {
     }
 
     fn try_sign_message(&self, message: &[u8]) -> Result<Signature, SignerError> {
-        match &self.wallet_type {
-            RemoteWalletType::Ledger(wallet) => wallet
-                .sign_message(&self.derivation_path, message)
-                .map_err(|e| e.into()),
-            RemoteWalletType::Trezor(wallet) => wallet
-                .sign_message(&self.derivation_path, message)
-                .map_err(|e| e.into()),
-        }
+        self.wallet_type
+            .sign_message(&self.derivation_path, message)
+            .map_err(|e| e.into())
     }
 
     fn is_interactive(&self) -> bool {
@@ -70,12 +61,16 @@ pub fn generate_remote_keypair(
     keypair_name: &str,
 ) -> Result<RemoteKeypair, RemoteWalletError> {
     let remote_wallet_info = RemoteWalletInfo::parse_locator(locator);
+
     let remote_wallet = get_wallet_from_info(remote_wallet_info, keypair_name, wallet_manager)?;
     let path = format!("{}{}", remote_wallet.path, derivation_path.get_query());
-    RemoteKeypair::new(
-        remote_wallet.wallet_type,
+    let wallet_type = remote_wallet.wallet_type;
+    let pubkey = wallet_type.get_pubkey(&derivation_path, confirm_key)?;
+
+    Ok(RemoteKeypair {
+        wallet_type,
         derivation_path,
-        confirm_key,
+        pubkey,
         path,
-    )
+    })
 }